requirements.txt
streamlit
alpaca-py
openpyxl
xlsxwriter